requirements.txt
Django==4.2.4
python-telegram-bot==13.15
django-phonenumber-field==7.1.0
phonenumbers==8.13.19
environs==9.5.0
Pillow==10.0.0
